backend/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
websockets
pyaudio
numpy
langchain==0.3.6
langchain-community>=0.3.5
langchain-openai
pypdf==5.1.0
openai-whisper
pydub
cached-path==1.6.3
coqui-tts>=0.24.3
simpleaudio>=1.0.4
pydub
passlib>=1.7.4 